feat: change Package name & add game_table_id + character_sheet_id
