--- src/main/java/com/poec/projet_backend/domaine/character/Character.java
@@ -1,21 +1,0 @@
-package com.poec.projet_backend.domaine.character;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import lombok.Data;
-import lombok.RequiredArgsConstructor;
-
-@Data
-@Entity
-@RequiredArgsConstructor
-public class Character {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    private String name;
-    private String avatar;
-    private boolean accepted;
-}
