--- src/main/java/com/poec/projet_backend/domaine/character/CharacterRepository.java
@@ -1,8 +1,0 @@
-package com.poec.projet_backend.domaine.character;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface CharacterRepository extends JpaRepository<Character, Long> {
-}
