--- src/main/java/com/poec/projet_backend/domaine/characterSheet/CharacterSheet.java
@@ -1,19 +1,0 @@
-package com.poec.projet_backend.domaine.characterSheet;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import lombok.Data;
-import lombok.RequiredArgsConstructor;
-
-@Data
-@Entity
-@RequiredArgsConstructor
-public class CharacterSheet {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-}
