--- src/main/java/com/poec/projet_backend/domaine/characterSheet/CharacterSheetService.java
@@ -1,4 +1,0 @@
-package com.poec.projet_backend.domaine.characterSheet;
-
-public class CharacterSheetService {
-}
