+++ src/main/java/com/poec/projet_backend/domaine/character_sheet/CharacterSheetController.java
@@ -1,0 +1,4 @@
+package com.poec.projet_backend.domaine.character_sheet;
+
+public class CharacterSheetController {
+}
