+++ src/main/java/com/poec/projet_backend/domaine/character_sheet/CharacterSheetService.java
@@ -1,0 +1,4 @@
+package com.poec.projet_backend.domaine.character_sheet;
+
+public class CharacterSheetService {
+}
