+++ src/main/java/com/poec/projet_backend/domaine/game_table/GameTableController.java
@@ -1,0 +1,4 @@
+package com.poec.projet_backend.domaine.game_table;
+
+public class GameTableController {
+}
