+++ src/main/java/com/poec/projet_backend/domaine/game_table/GameTableRepository.java
@@ -1,0 +1,4 @@
+package com.poec.projet_backend.domaine.game_table;
+
+public interface GameTableRepository {
+}
