+++ src/main/java/com/poec/projet_backend/domaine/game_table/GameTableService.java
@@ -1,0 +1,4 @@
+package com.poec.projet_backend.domaine.game_table;
+
+public class GameTableService {
+}
