+++ src/main/java/com/poec/projet_backend/domaine/playerCharacter/PlayerCharacterController.java
@@ -1,0 +1,4 @@
+package com.poec.projet_backend.domaine.playerCharacter;
+
+public class PlayerCharacterController {
+}
