+++ src/main/java/com/poec/projet_backend/domaine/playerCharacter/PlayerCharacterRepository.java
@@ -1,0 +1,4 @@
+package com.poec.projet_backend.domaine.playerCharacter;
+
+public interface PlayerCharacterRepository {
+}
