+++ src/main/java/com/poec/projet_backend/domaine/playerCharacter/PlayerCharacterService.java
@@ -1,0 +1,4 @@
+package com.poec.projet_backend.domaine.playerCharacter;
+
+public class PlayerCharacterService {
+}
