--- src/main/java/com/poec/projet_backend/domaine/table/Table.java
@@ -1,20 +1,0 @@
-package com.poec.projet_backend.domaine.table;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import lombok.Data;
-import lombok.RequiredArgsConstructor;
-
-@Data
-@Entity
-@RequiredArgsConstructor
-public class Table {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    private String avatar;
-    private String name;
-}
