--- src/main/java/com/poec/projet_backend/domaine/table/TableController.java
@@ -1,4 +1,0 @@
-package com.poec.projet_backend.domaine.table;
-
-public class TableController {
-}
