--- src/main/java/com/poec/projet_backend/domaine/table/TableRepository.java
@@ -1,4 +1,0 @@
-package com.poec.projet_backend.domaine.table;
-
-public interface TableRepository {
-}
