--- src/main/java/com/poec/projet_backend/domaine/table/TableService.java
@@ -1,4 +1,0 @@
-package com.poec.projet_backend.domaine.table;
-
-public class TableService {
-}
